Funcionalidad de búsqueda hecha
-Limpieza del código

+++ app/Http/Controllers/TaskController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\{Task, User};
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
 use Illuminate\Http\Request;
 
 class TaskController extends Controller
@@ -15,7 +17,19 @@ class TaskController extends Controller
      */
     public function index()
     {
-        return view('tasks');
+        //Admin and User allowed
+
+        //If is admin get all tasks
+        if(Gate::allows('view_all_tasks_access')){
+            $tasks = Task::with(['creator_user','assigned_user'])->get();
+        }
+        else {
+            //If is user get user's tasks
+            $loggedUserId = Auth::id();
+            $tasks = Task::with(['assigned_user', 'creator_user'])->where('assigned_user_id',$loggedUserId)->get();
+        }
+
+        return view('tasks.index', compact('tasks'));
     }
 
     /**
@@ -25,13 +39,15 @@ class TaskController extends Controller
      */
     public function create()
     {
+        //Admin and User allowed
+
         //Declaration of vars in order to use the same form template
         $task = new Task;
         $title = "Nueva Tarea";
         $txtButton = "Agregar";
         $route = route('tasks.store');
         $users = User::all();
-        return view('create-task', compact('task','title','txtButton','route','users'));
+        return view('tasks.create', compact('task','title','txtButton','route','users'));
     }
 
     /**
@@ -42,6 +58,8 @@ class TaskController extends Controller
      */
     public function store(Request $request)
     {
+        //Admin and User allowed
+
         $validated = $request->validate([
             'name' => 'required|max:100',
             'start_date' => 'required|date|after_or_equal:yesterday',
@@ -57,13 +75,22 @@ class TaskController extends Controller
         //Save related users to task
         $loggedUserId = Auth::id();
         $creator_user = User::find($loggedUserId);
-        $assigned_user = User::find($request->assigned_user);
+        
+        //Check if the user is allowed to assign users to task
+        if(Gate::allows('assigned_user_access')){
+            $assigned_user = User::find($request->assigned_user);
+        }
+        else {
+            //if not, creator user is assign to own tasks
+            $assigned_user = $creator_user;
+        }
+        
         $task->creator_user()->associate($creator_user);
         $task->assigned_user()->associate($assigned_user);
 
         $task->save();
 
-        return view('tasks');
+        return redirect()->route('tasks.index');
     }
 
     /**
@@ -74,21 +101,25 @@ class TaskController extends Controller
      */
     public function edit(Task $task)
     {
+        //Admin allow only 
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         //Declaration of vars in order to use the same form template
         $update = true;
         $title = "Actualizar Tarea";
         $txtButton = "Actualizar";
         $route = route('tasks.update',['task' => $task]);
         $users = User::all();
-        return view('edit-task', compact('task','title','txtButton','route','users','update'));
+        return view('tasks.edit', compact('task','title','txtButton','route','users','update'));
     }
 
-    //Get the specified resource
-    public function getTask($id)
+    //Mark task as completed
+    public function completeTask(Task $task)
     {
-        $task = Task::find($id);
+        $task->done = 1;
+        $task->save();
 
-        return response()->json(['status' => 'success', 'task' => 'task']);
+        return back();
     }
 
     /**
@@ -100,6 +131,9 @@ class TaskController extends Controller
      */
     public function update(Request $request, Task $task)
     {
+        //Admin allow only 
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $this->validate($request, [
             'name' => 'required|max:100|unique:tasks,name,' . $task->id,
             'start_date' => 'required|date|after_or_equal:yesterday',
@@ -117,7 +151,7 @@ class TaskController extends Controller
 
         $task->save();
 
-        return view('tasks');
+        return redirect()->route('tasks.index');
     }
 
     /**
@@ -128,8 +162,35 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        //Admin allow only 
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $task->delete();
 
         return back();
+    }
+
+    public function search(Request $request)
+    {
+        $this->validate($request, [
+            'search_criteria' => 'string|max:100'
+        ]);
+
+        //If is admin get all tasks
+        if(Gate::allows('view_all_tasks_access')){
+            $tasks = Task::with(['creator_user','assigned_user'])
+                ->where('name', 'like', '%'.$request->search_criteria.'%')
+                ->get();
+        }
+        else {
+            //If is user get user's tasks
+            $loggedUserId = Auth::id();
+            $tasks = Task::with(['assigned_user', 'creator_user'])
+                ->where('assigned_user_id',$loggedUserId)
+                ->where('name', 'like', '%'.$request->search_criteria.'%')
+                ->get();
+        }
+
+        return view('tasks.index', compact('tasks'));
     }
 }
